refactor: changed internal APIs bug: nested statements in components
Now using more references for individual module an export getters for
example.

There is still a pending bug to fix which is getting a service method
usage in nested statements in a component's methods. For example, if a
method service is used inside a nested if statement, this is not
registered as a valid usage right now, but it should.