feat: cli client and frontend scraper entrypoint
Started building the CLI client for scraping the frontend (still missing
getting the actual constant from the CLI and error handling of course).

I have to redefined the spec for a frontend dir and make all of the
three files (constants, service, component) required.

Also have to map a constant's name to it's service method identifier
instead of the whole signature (not map it directly since I care about
other info in the method).