refactor(cli): removed search query params impl
It is better to used a better tool like jq. This will give the user much
more flexibility when searching for a constant. The scraper indexes all
the project anyways so its not like I'm gaining time finishing the
search after a match if I implement this directly into the program.